+++ app/src/main/java/com/mrdarip/stepy/domain/model/Execution.kt
@@ -1,10 +1,10 @@
 package com.mrdarip.stepy.domain.model
 
 data class Execution(
-    val id: Int = 0,
-    val parentRoutineId: Int?,
-    val parentExecutionId: Int?,
-    val stepId: Int,
+    val id: Long = 0,
+    val parentRoutineId: Long?,
+    val parentExecutionId: Long?,
+    val stepId: Long,
     val position: Int,
     val start: Long,
     val end: Long
